demo/requirements.txt
amqp==1.4.6
anyjson==0.3.3
backports.shutil-get-terminal-size==1.0.0
billiard==3.3.0.20
celery==3.1.18
decorator==4.0.11
Django==1.8.3
django-activity-stream==0.6.0
django-appconf==1.0.1
django-bootstrap-breadcrumbs==0.7.1
django-celery==3.1.16
django-cors-headers==1.1.0
django-datatables-view==1.12.1
django-debug-panel==0.8.1
django-debug-toolbar==1.3.2
django-debug-toolbar-line-profiler==0.4.0
django-extensions==1.5.5
django-jsonfield==1.0.1
django-jsonify==0.3.0
django-mptt==0.7.4
django-passwords==0.3.3
django-realtime==1.1
django-redis==4.2.0
django-redis-cache==1.1.1
django-rest-swagger==0.3.2
django-session-security==2.2.4
django-settings-export==1.0.5
django-staticfiles==1.2.1
djangorestframework==3.1.3
enum34==1.1.6
hiredis==0.2.0
ipaddr==2.1.11
ipdb==0.10.2
IPy==0.83
ipython==5.3.0
ipython-genutils==0.2.0
kombu==3.0.26
line-profiler==1.0
memory-profiler==0.33
msgpack-python==0.4.6
MySQL-python==1.2.5
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==2.9.0
prompt-toolkit==1.0.13
ptyprocess==0.5.1
pydot==1.0.2
pyenchant==1.6.6
Pygments==2.2.0
pymongo==3.0.3
pyparsing==2.0.3
pyproj==1.9.4
python-dateutil==2.4.2
python-memcached==1.54
pytz==2015.4
PyYAML==3.11
raven==5.4.4
redis==2.10.3
requests==2.7.0
scandir==1.5
Shapely==1.5.9
simplegeneric==0.8.1
simplejson==3.7.3
six==1.9.0
sqlparse==0.1.15
suds==0.4
traitlets==4.3.2
ujson==1.33
uWSGI==2.0.11
wcwidth==0.1.7
xlrd==0.9.3
XlsxWriter==0.8.4
xlwt==1.0.0